requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
ollama==0.4.4
python-dotenv==1.0.1
pydantic==2.10.4
